% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EoApostL.R
\name{EoApostL}
\alias{EoApostL}
\alias{EoApostL_parameters}
\title{Process EoA Lambda Posteriors}
\usage{
EoApostL()

EoApostL_parameters(X = NULL, Ba = NULL, Bb = NULL)
}
\value{
\code{list} of posteriors of lambda.
}
\description{
Collect and process Lambda posteriors from EoA.
  \cr
  Run this function after conducting a single year analysis.
}
