% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EoApostM.R
\name{EoApostM}
\alias{EoApostM}
\title{Posterior of M}
\usage{
EoApostM(posterior = "FSFY", PCIp = c(0.025, 0.975), digits = c(5, 2), ...)
}
\arguments{
\item{posterior}{Which of the three posteriors}

\item{PCIp}{posterior credibility interval of interest}

\item{digits}{decimal places to leave}

\item{...}{...}
}
\value{
\code{list} of EmpiricalPosterior and SummaryStats
}
\description{
Collect and process the posterior M from EoA.
}
