% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{overlap}
\alias{overlap}
\title{Measure overlap between pairs of CDFs}
\usage{
overlap(Fs, vals, digits = 10, ...)
}
\arguments{
\item{Fs}{\code{list} of CDFs}

\item{vals}{\code{numeric} vector for where \code{Fs} should be 
differentiated}

\item{digits}{rounding digits}

\item{...}{additional arguments to pass to \code{\link[pracma]{numdiff}}}
}
\value{
\code{matrix} of pairwise overlap values
}
\description{
Measures overlap between pairs of CDFs. First uses
 \code{\link[pracma]{numdiff}} to differentiated the CDFs, then calculates
 the classic Overlap Coefficient.
}
